+++ src/PlotBoilerplate.js
@@ -66,7 +66,7 @@
     var setCSSscale = function( element, scaleX, scaleY ) {
 	element.style['transform-origin'] = '0 0';
 	if( scaleX==1.0 && scaleY==1.0 ) element.style.transform = null;
-	else                             element.style.transform = 'scale(' + scaleX + ',' + scaleY+')';
+	else                             element.style.transform = 'scale(' + scaleX + ',' + scaleY + ')';
     };
 
 
@@ -156,6 +156,7 @@
 	    fitToParent           : typeof config.fitToParent != 'undefined' ? config.fitToParent : true,
 	    scaleX                : config.scaleX || 1.0,
 	    scaleY                : config.scaleY || 1.0,
+	    drawGrid              : typeof config.drawGrid != 'undefined' ? config.drawGrid : true,
 	    rasterGrid            : typeof config.rasterGrid != 'undefined' ? config.rasterGrid : true,
 	    rasterAdjustFactor    : typeof config.rasterAdjustFactor == 'number' ? config.rasterAdjustFactor : 2.0,
 	    drawOrigin            : typeof config.drawOrigin != 'undefined' ? config.drawOrigin : false,
@@ -376,10 +377,12 @@
 	    var offset = this.draw.offset.clone().inv();
 	    offset.x = (Math.round(offset.x+cs.w)/Math.round(gSize.w))*(gSize.w)/this.draw.scale.x + (((this.draw.offset.x-cs.w)/this.draw.scale.x)%gSize.w);
 	    offset.y = (Math.round(offset.y+cs.h)/Math.round(gSize.h))*(gSize.h)/this.draw.scale.y + (((this.draw.offset.y-cs.h)/this.draw.scale.x)%gSize.h);
-	    if( this.config.rasterGrid )
-		this.draw.raster( offset, (this.canvasSize.width)/this.draw.scale.x, (this.canvasSize.height)/this.draw.scale.y, gSize.w, gSize.h, 'rgba(0,128,255,0.125)' );
-	    else
-		this.draw.grid( offset, (this.canvasSize.width)/this.draw.scale.x, (this.canvasSize.height)/this.draw.scale.y, gSize.w, gSize.h, 'rgba(0,128,255,0.095)' )
+	    if( this.config.drawGrid ) {
+		if( this.config.rasterGrid )
+		    this.draw.raster( offset, (this.canvasSize.width)/this.draw.scale.x, (this.canvasSize.height)/this.draw.scale.y, gSize.w, gSize.h, 'rgba(0,128,255,0.125)' );
+		else
+		    this.draw.grid( offset, (this.canvasSize.width)/this.draw.scale.x, (this.canvasSize.height)/this.draw.scale.y, gSize.w, gSize.h, 'rgba(0,128,255,0.095)' );
+	    }
 	};
 
 	
@@ -618,6 +621,25 @@
 	};
 
 
+	/**
+	 * Get the current view port.
+	 *
+	 * @method viewPort
+	 * @instance
+	 * @memberof PlotBoilerplate
+	 * @return {Bounds} The current viewport.
+	 **/
+	PlotBoilerplate.prototype.viewport = function() {
+	    return { min : this.transformMousePosition(0,0),
+		     max : this.transformMousePosition(this.canvasSize.width,this.canvasSize.height)
+		   };
+	};
+	/**
+	 * @typedef {Object} Bounds
+	 * @property {Vertex*} min The upper left position.
+	 * @property {Vertex*} max The lower right position;.
+	 */
+
 
 	/**
 	 * This function opens a save-as file dialog and – once an output file is
@@ -852,7 +874,7 @@
 	    if( e.which != 1 && !(window.TouchEvent && e.originalEvent instanceof TouchEvent) )
 		return; // Only react on left mouse or touch events
 	    var p = locatePointNear( _self.transformMousePosition(e.params.pos.x, e.params.pos.y), DEFAULT_CLICK_TOLERANCE/Math.min(_self.config.cssScaleX,_self.config.cssScaleY) );
-	    _self.console.log('point at position found: '+p );
+	    //_self.console.log('point at position found: '+p );
 	    if( !p ) return;
 	    // Drag all selected elements?
 	    if( p.type == 'vertex' && _self.vertices[p.vindex].attr.isSelected ) {